flask/requirements.txt
flask==3.1.0
langgraph==0.2.45
google-genai==0.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.3